fixed mapping error

+++ src/main/java/com/tngovschools/admission_system/model/Student.java
@@ -11,7 +11,7 @@ import lombok.*;
 public class Student {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long rollNo;
+    private int rollNo;
 
     private String name;
     private String fatherName;
